Factor out dtype for networks

+++ sloika/features.py
@@ -1,5 +1,5 @@
 import numpy as np
-from sloika import layers
+from sloika import sloika_dtype, layers
 from sloika.maths import studentise
 
 def from_events(ev):
@@ -9,12 +9,12 @@ def from_events(ev):
     :Returns: A :class:ndarray with studentised features
     """
     nev = len(ev)
-    features = np.zeros((nev, 4), dtype=layers.dtype)
+    features = np.zeros((nev, 4), dtype=sloika_dtype)
     features[:,0] = ev['mean']
     features[:,1] = ev['stdv']
     features[:,2] = ev['length']
     #  Zero pad delta mean
     features[:,3] = np.fabs(np.ediff1d(ev['mean'], to_end=0))
 
-    features = np.ascontiguousarray(studentise(features, axis=0), dtype=layers.dtype)
+    features = np.ascontiguousarray(studentise(features, axis=0), dtype=sloika_dtype)
     return features
